+++ src/main/java/com/atlassian/performance/tools/sshubuntu/api/SshUbuntu.java
@@ -1,5 +1,0 @@
-package com.atlassian.performance.tools.sshubuntu.api;
-
-public interface SshUbuntu extends AutoCloseable {
-    SshHost getSsh();
-}