+++ src/main/kotlin/com/atlassian/performance/tools/sshubuntu/api/SshUbuntuImage.kt
@@ -1,0 +1,7 @@
+package com.atlassian.performance.tools.sshubuntu.api
+
+interface SshUbuntuImage {
+    fun <T> runInUbuntu(
+        lambda: (SshUbuntuContainer) -> T
+    ): T
+}