+++ CHANGELOG.md
@@ -23,6 +23,15 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ## [Unreleased]
 [Unreleased]: https://github.com/atlassian/ssh-ubuntu/compare/master...release-0.1.0
 
+### Removed
+- Remove `SshUbuntu`.
+
+### Added
+- Add `SudoSshUbuntuImage`. Resolve [JPERF-445].
+- Add `AutoCloseable` wrappers for some `docker-java` resources.
+
+[JPERF-445]: https://ecosystem.atlassian.net/browse/JPERF-445
+
 ## [0.1.0] - 2019-02-15
 [0.1.0]: https://github.com/atlassian/ssh-ubuntu/compare/release-0.1.0...initial-commit