+++ CHANGELOG.md
@@ -21,7 +21,7 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://github.com/atlassian/ssh-ubuntu/compare/master...release-0.1.0
+[Unreleased]: https://github.com/atlassian/ssh-ubuntu/compare/release-0.1.0...master
 
 ### Added
 - Customize `SshUbuntuContainer` via `testcontainers` API. Resolve [JPERF-445].
@@ -29,7 +29,7 @@ Dropping a requirement of a major version of a dependency is a new contract.
 [JPERF-445]: https://ecosystem.atlassian.net/browse/JPERF-445
 
 ## [0.1.0] - 2019-02-15
-[0.1.0]: https://github.com/atlassian/ssh-ubuntu/compare/release-0.1.0...initial-commit
+[0.1.0]: https://github.com/atlassian/ssh-ubuntu/compare/initial-commit...release-0.1.0
 
 ### Added
 - Provide `SshUbuntu`.